Simplify sum through many cache by adding some constraints
The combination of foreign keys in the join table must now be
unique, as the trigger on the associated table does not handle
the update correctly if there are duplicate rows.  While the
trigger function could be changed to account for that, I think
it's better to keep things simple, since I doubt anyone needs
to handle duplicates.

Require the certain columns be NOT NULL, so the trigger functions
do not need to check for that.

Require id columns be primary keys, which must be unique and not
null, and avoid the aggregate in the join table trigger function.

Avoid an empty IF clause in the join table trigger function by
inverting the condition.

Add some additional specs.