Add a pg_triggers Sequel extension only affecting a single database
The previous way of requiring the library directly and having it
make a global modification still works, but is now discouraged.
The implementation is kind of clunky.  Since we can't include the
extension module directly in Sequel::Postgres::DatabaseMethods, as
that only affects future includes of
Sequel::Postgres::DatabaseMethods, add most of the method definitions
to a proc, and class_eval the proc both in the global modification
and in the extension module.

This probably hides the method RDoc, so I'll move that into the
README.

Remove some leftover RDoc cruft while here.