Make counter and sum caches handle INSERT ... ON CONFLICT DO NOTHING (Fixes #15)
Note that if you already have defined counter or sum cache triggers
prior to this change, and want them to work correctly when using
INSERT ... ON CONFLICT DO NOTHING, you should drop the triggers and
then readd them.

The reason this wasn't handled originally was that these triggers
were originally developed in 2008, and ON CONFLICT support was not
added to PostgreSQL until 2016.